Set cloud provider for cluster components (#1386)
